fix build, rebuild, make clean file exist

+++ Makefile
@@ -1,12 +1,16 @@
 build: extract
-	mv export/post/* site/content/post/
-	mv export/media/* site/content/media/
+	@echo hugo build site to public/
+	@if [ -e export/post/* ]; then mv export/post/* site/content/post/ ; fi
+	@if [ -e export/media/* ]; then mv export/media/* site/content/media/ ; fi
+	@if [ -e public ]; then rm -r public ; fi
 	@cd site ; \
 	hugo ; \
 	cd .. ; \
 	mv site/public ./
 
 extract: builder
+	@echo extract from Notion.so
+	@if [ -e export ]; then rm -r export ; fi
 	@builder/builder
 
 builder: mod fmt
@@ -26,6 +30,6 @@ mod:
 
 .PHONY: clean
 clean:
-	rm builder/builder
-	rm -r export
-	rm -r public
+	if [ -e builder/builder ]; then rm builder/builder ; fi
+	if [ -e export ]; then rm -r export ; fi
+	if [ -e public ]; then rm -r public ; fi
